requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.9.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
joblib==1.1.0
numpy==1.21.2
oauthlib==3.1.1
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.2.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.2.0
python-dateutil==1.5
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sqlparse==0.4.1
threadpoolctl==3.0.0
urllib3==1.26.7
virtualenv==20.7.2
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.6
whitenoise==5.3.0
